requirements.txt
flask
sentence-transformers
torch
torchvision
faiss-cpu
Pillow
fitz
tools
fronend
numpy
pandas
pymupdf # Check the correctness of this
